FISH-318 Move SNMP Notifier Package
