+++ appserver/payara-appserver-modules/notification-snmp-core/src/main/java/fish/payara/extras/notifiers/snmp/SetSnmpNotifierConfigurationCommand.java
@@ -35,7 +35,7 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package fish.payara.notification.snmp;
+package fish.payara.extras.notifiers.snmp;
 
 import java.beans.PropertyVetoException;
